Adding `vm.br_table` op. (#15286)
This implements a 0-based dense branch table that can be lowered from
`cf.switch` which is a weird op but what we have and what comes from
`scf.index_switch`. The exact encoding may change in the future to make
interpreting faster but as structured it should be good for basic JITs.

The initial usage of this will come from entirely dense (or mostly
dense) `scf.index_switch` ops so only the most basic optimization for
offseting the base of the table is implemented. There's no
canonicalization/folding as `cf.switch` has most of that but we should
probably add some of the same handlers in the future.