Adding selection condition to hal.executable.variant. (#15284)
This allows for variants to declare host logic that determines whether
the variant should be selected for loading. When multiple variants are
available their declared conditions will be evaluated in op order along
with the existing executable format match.

Unfortunately the MLIR SymbolTable trait disallows multiple regions on
any op holding it so a new `hal.executable.condition` region op was
added that may be optionally present on any `hal.executable.variant`.
Ideally we clean this up and make it an optional region but that'll need
relaxing of upstream assertions like
https://sourcegraph.com/github.com/llvm/llvm-project/-/blob/mlir/lib/IR/SymbolTable.cpp?L122-123
(ideally either treating region 0 as the symbol table on ops or having
an interface override for selecting the region ala `getCallableRegion`
such as `getSymbolTableRegion`).

This removes the `hal.device.switch` op in favor of `scf.index_switch`.
When we start running const expr hoisting during the HAL pipeline this
should allow variant selection to be completely hoisted to
initialization time (or at least memoized per device). There's decent
low-hanging future work on optimizing the ranking/selection and
improving `scf.index_switch` hoisting/canonicalization to make things
better.