Cast tensor.empty type to TypeConverter's type during materialization. (#15375)
It is a bug introduced by
https://github.com/openxla/iree/commit/ab9e7b702dbabab05097611b8e47c491167cb8bd

VMVX has a different assumption because it generally supports dynamic
cases. I.e., it treats all the matmul as dynamic shapes even they are
static inputs. This could make inferred shape and TypeConverter's target
shape different in terms of dynamism. Casting it to the target type to
make dialect conversion happy.