Adding support for splat entries in parameter indices. (#15420)
Parameters can now either be backed by a file or a splat pattern. Splat
parameters are read-only as no storage is reserved for them. This PR
adds the handling of different entry types and allows parameter index
users to declare them but does not yet add support for producing these
via in-tree tools.

Progress on #14987.