[CPU] Improve tile sizes selection for tensor.pack ops. (#15397)
It disables special vector sizes for non-f32 cases because the logic is
only for 16x16 transpose cases. The improvements of dispatch sizes are
from vectorization. We are not able to vectorize named ops if they have
dynamic shapes, which is fixed by
https://github.com/llvm/llvm-project/commit/03529b99b36788ca836b7ce238ea9400ce89847b.
The change allows backends to vectorize them because they become static
shapes (by tiling with size=1). It is not a hard condition; we track it
in https://github.com/openxla/iree/issues/15441

The revision takes the number of threads into account, so we have better
performance on multi-threaded. It also reduces runtime overheads.

This is a step toward to https://github.com/openxla/iree/issues/15391
and https://github.com/openxla/iree/issues/15349

It improves the performance of
[tensor.pack](https://github.com/openxla/iree/issues/15349) op from 420
ms to 170 ms on 8-threaded x86 CPU.