Fix `fp16` feature on arm64: the proper feature name is `fullfp16`, not `fp16`. (#15479)
The symptom being fixed here is these warnings while building
`iree-test-deps` on arm64:

```
'+fp16' is not a recognized feature for this target (ignoring feature)
```

This was saying that bytecode modules that are meant to be compiled with
fp16 native arithmetic instructions were actually compiled without them,
falling back on slower paths converting to fp32.

The problem was that we were mixing up two concepts in the LLVM AArch64
target: "extension" vs "feature". The features are the individual,
mutually orthogonal things that we want to be working with here. The
feature name here is `fullfp16`. Extensions are a higher-level
construct, a group of features. Many extensions contain only one
feature, and in that case, most of the time, the extension name equals
the feature name. But `fp16` is an exception:


https://github.com/llvm/llvm-project/blob/26ab444e88fc8fdd554e5a9381a68b7b5e63b6fd/llvm/lib/Target/AArch64/AsmParser/AArch64AsmParser.cpp#L3639

I think the most important dimension here is orthogonality. Features are
mutually orthogonal, so they are what we want to reflect in our own CPU
feature bits.

The one place where we need to keep working with extensions is in Clang
copts, when building either native code or bitcode in ukernels. There,
the `-march=armv8.2-a+...` wants extensions, not features.