[HAL] Add configuration stage in executable lowerings (#15362)
The core idea behind this split is to allow for passes to be added in
between strategy selection and translation, thereby expanding and
contracting the set of variants per executable based on the known
feature requirements for those strategies.

The reason a split is needed is because strategy selection itself should
not care what other variants of the executable are being compiled,
however to deduplicate them we need to nest at the executable level.
Codegen was already partially structured to enable this (by way of
encoding the lowering pipelines as an IR attribute). This PR just
recovers the separation that was already there.

In terms of immediate benefits, this adds an option to dump executable
configurations (essentially post-kernel config dispatches) and
substitute them in, making it significantly easier to get the default
lowering configuration for a dispatch from the compiler, tweak it, and
plug it back in. Additionally benchmark dumps now happen after this
stage for the same reason.