[GlobalOptimization] Support ExpandVectors matvec cases with producer CastOpInterface ops (#15372)
We currently lift `linalg.generic` ops into `linalg.batch_matvec` with
`CastOpInterface` ops as separate producer ops, instead of creating a
matvec with implicit element type promotion. We need to keep the
`CastOpInterface` ops as the producer of the `linalg.batch_matmul` ops
in order to recognize this sequence of ops as a matmul with casted types
in SetEncoding.