Optimized vecmat ukernel tile functions for `i16 x u4 -> i32` on AVX-512-VNNI (#15525)
This kernel is parametrized in N0, allowing N0==16 and N0==32.
Performance on AMD Ryzen 9 7950X3D:
  - With N0=16:  180 Gop/s.
  - With N0=32:  240 Gop/s.

These numbers show that there's a nice reward for going extra large, but
that's also a liability for vecmat shapes whose N dimension isn't a
multiple of 32. Maybe we can keep both for now.

This is currently by far our fastest vecmat tile function --- it's fast
even by general-matmul standards, while usually vecmat's low arithmetic
intensity relegates it to lower performance levels. It shows what's
possible now that we've decoupled vecmat tile shapes from general matmul
tile shapes in #15431 . That 32x8 is not a truncation of a general
matmul tile shape. Other element types and CPU architectures all need to
get the same treatment.

The idea of this kernel is to split the LHS s16 values into high and low
8-bit components to be able to use `_mm512_dpbusd_epi32`.

In itself, that doesn't reduce the number of arithmetic instructions:
while each now computes a 4D dot-product instead of a 2D one as in
`_mm512_dpwssd_epi32`, we now need twice more of them to do separately
the high and low 8bit parts of the LHS s16 values.

The real benefit is that this removes the need to extend RHS u4 values
to s16. Since this is a vecmat kernel, the LHS is small and the RHS is
big, so it matters to avoid RHS-processing work.

It's not trivial how to use `_mm512_dpbusd_epi32`, with its quirky
unsigned * signed semantics. We take advantage of the fact that our u4
RHS values, when extended to u8, do not use the top bit -- so they are
also interpretable as s8 values in place. So this is specific to RHS
being less-than-8-bit values (it's not specific beyond that to 4bit).
Meanwhile, when we split the LHS s16 values into high and low 8bit
components the high 8bits are signed s8 and the low 8bit are unsigned
u8. So, for each of the combinations of operands that we have to feed
`_mm512_dpbusd_epi32`, we manage to find an operand order that
accomodates the instruction's requirements on signednesses.