Delete IREEImportPublicTest. (#15573)
I'm not fully sure what this test is trying to add coverage for
(`IREE::HAL::getElementTypeValue` matches
`IREE::Input::getElementTypeValue`...?), but this isn't the place for
that type of testing (maybe static asserts or tests using the compiler
API instead?). We do already have
[`test/iree_import_public.mlir`](https://github.com/openxla/iree/blob/main/compiler/src/iree/compiler/InputConversion/Common/test/iree_import_public.mlir`).

Practically speaking, this test got in the way of some refactoring on
https://github.com/openxla/iree/pull/15568 (that PR starts generating
this CMakeLists.txt from the associated BUILD.bazel file and moves some
dependencies around).