`--iree-llvmcpu-enable-ukernels`, a fine-grained replacement for `--iree-llvmcpu-enable-microkernels` (#15571)
This introduces a new command-line flag,
`--iree-llvmcpu-enable-ukernels`, as a replacement for the existing
`--iree-llvmcpu-enable-microkernels`, which gets deprecated but
temporarily coexists during the transition.

The name ambiguity (only differing by `micro` vs `u` substring) is OK as
that's part of a general renaming trend and the old flag should get
removed quickly.

The problem with the boolean `--iree-llvmcpu-enable-microkernels` was
that our plan was of course to eventually enable it by default, but then
that puts a very high bar on any future new ukernel. For now, the only
ukernel enabled in LLVMCPU is mmt4d, but the next time that we want to
introduce a new ukernel, if that means that all existing codegen users
get it while it is still under development, that would be hard.

The new `--iree-llvmcpu-enable-ukernels` is a string. It has the default
value `default`, which means let the implementation enable whichever set
of ukernels is currently default. Currently, that's no ukernel at all,
but we hope in the near team to change that default to including the
`mmt4d` ukernel. Other special string values are `none` and `all` with
the obvious meanings. Any other string is interpreted as a
comma-separated list of individual ukernels to enable.

The old boolean flag `--iree-llvmcpu-enable-microkernels={false,true}`
is currently interpreted as `--iree-llvmcpu-enable-ukernels={none,all}`
until it's dropped.