Drop `--iree-llvmcpu-enable-microkernels`, replaced by `--iree-llvmcpu-enable-ukernels` (#15646)
This old flag (`-microkernels`) was a boolean. The new flag
(`-ukernels`) is a string.

Porting guide:
- If you were passing `--iree-llvmcpu-enable-microkernels` (or
equivalently, `...=true`), the equivalent new flag is
`--iree-llvmcpu-enable-ukernels=all`. You may however choose to do
`--iree-llvmcpu-enable-ukernels=mmt4d` instead if you want to only
enable the mmt4d ukernel and not all future ukernels that may be added
in the future.
- If you were passing `--iree-llvmcpu-enable-microkernels=false`, that
was the default behavior anyway so that wasn't having any effect, so you
could drop that; however, if you meant to explicitly opt out of ukernels
whenever they would become default, the equivalent of that with the new
flag is `--iree-llvmcpu-enable-ukernels=none`.

All occurences of the old flag in the IREE repo have already been
replaced. Please confirm when it's OK to proceed with this, @mariecwhite
@pzread @dcaballe @jpienaar @lundong @hcindyl @banach-space .