[CPU][NFC] Unify the strategy selection for TilingInterface and LinalgOp (#15774)
The two entries use the same logic. The TilingInterface method extracts
`lbs` and `rbs` through `.getIterationDomain()` method; the LinalgOp
method extracts the information through `.getStaticLoopRanges()`. The
former one is more generic than the latter one. We can use the former
one to get what we want. This revision removes the entry for LinlagOp
and leverage the logic to TilingInterface's method.