Adding condition support to `hal.dispatch.extern`. (#15797)
Required switching from `hal.dispatch.extern` outlining into a
`flow.executable` to instead outlining into a `hal.executable` with
variants. This is nice as it prevents us from needing to carry through
all the information a `hal.executable` could have through flow/stream
but does mean we need to support `flow.dispatch`/`stream.async.dispatch`
on expanded executables with variants that may have multiple export
symbols. In the common case (codegen) there's no change.

It's not great this is happening in flow - future work may try to move
outlining of the extern ops to the end of the global optimization
pipeline (assuming that happens _after_ plugins that may insert the
extern ops). There's a few refactorings in other passes we'd need to
make that happen.