Delete old integrate scripts and add export_submodule_head.py. (#15846)
The old integrate scripts have been superceded by megabump, which is
maintained out of tree for the moment.

As it is of general use, megabump's export_submodule_head.py script is
added directly to IREE so it can be used by both the integrate process
and normal developer workflow for publishing submodule cherrypicks.

This script has evolved over a very long time, so look kindly on the
style. I basically inlined some utilities that had accumulated out of
tree and rolled into one thing.