[spirv] Respect entry point ordinal when serializing executables (#15905)
We have assigned ordinals to various entry points when linking; need to
respect the order there when serializing. We were lucky before because
hal.executable.export ops are sorted in ascending order w.r.t. ordinals
thus far, and spirv.module ops follow the same order of
hal.executable.export ops. But it's not guaranteed to be so.