[LLVMGPU] Re-enable multidim distribution on gridDimension. (#15924)
When running larger model such as Llama2, we will come across matvecs or
matvec like operations with a large batch or parallel dimensions. Since
we tile the batch/parallel dim with size 1, and lay it across the grid,
we may hit into gridDim overflow on a single axis. This PR re-enables
multidim grid distribution to fix that issue.