[spirv] Trim provided executable target before linking (#15888)
This commit adds a pass to trim the provided target environment to the
minimal required one to run the compiled spirv.module op. This increases
the chance of linking different executable variants together when giving
multiple targets. For the same provided target, it may actually generate
more exectuables, given that now each dispatch may see a different set
of capability requirements. So we may need to be smarter when linking
executables--simply relying on exact target attribute match, which is
the current state, is not the best way. But this pass alone should be a
good cleanup.