[GlobalOptimization] Add pass for transpose propagation (#15858)
This adds a pass for propagating transposes already present in the
model, or those that were introduced in earlier passes, before data
tiling. The important note with this pass is that it does not do any
real analysis. Data tiling is the de facto place where layout decisions
are made, and similarly is the place where we would want to universally
propagate such layouts (and potentially do an analysis). This pass is
intended just to do local propagation of transposes only in cases that
are universally lucrative, or resolvable by data tiling.

With that note, the idea is that the passes/patterns themselves should
be relatively simple. It is organized into 4 phases. 1) Specialize all
transpose-like ops to linalg.transpose. This makes the
   propagation patterns significantly easier to write.
2) Sink transposes down. A few of the patterns for this are described
   below.
[2.5) We will want to add bubbling patterns as the need arises, but that
      only run for transposes we don't mark as "terminal" after sinking]
3) Combine adjacent transposes.
4) Re-generalize all transposes because the pipeline expects it.

The currently implemented propagation patterns are:

- Sinking through tensor.expand_shape. This is always free and helps
reveal new propagation opportunities
- Sinking through tensor.extract_slice. Currently this restricts to
cases where the resulting transposed slice is known to be contiguous,
however we could choose to always do this as a transpose on a slice is
likely to be cheaper than on the whole thing.
 - Named op -> named op conversions
    - linalg.matmul <-> linalg.matmul_transpose_b/matmul_transpose_a
    - linalg.batch_matmul <-> linalg.bmm_transpose_b/bmm_transpose_a
- Generic input fusion. This simply does an in-place rewrite of the
generic by transposing the results of the associated input's indexing
map.

Additionally a flag for aggressive transpose propagation is added that
allows for named op -> generic rewrites when there is no equivalent
named op to rewrite to. In the fullness of time, this should be turned
on by default, but there are places in the code that currently expect
certain named ops.