[VM] Add a vm.rodata.table.inline op (#15971)
Tables are a vm pseudo op for flattened data indexed by subviews stored
as independent read-only data as an array of (offset, size) values. This
op is expected to be decomposed into two vm.rodata.inline ops.

There might be room to add another non-inline version of this op and do
global program analysis to pack other read only data into a single table
with a set of views.