[spirv] Provide same entry point set in variants when linking (#15935)
Note that at runtime, for a particular executable, only one variant of
it will be loaded. So, all variants of an executable are expected to
provide the exact same set of entry points; this way we can guarantee no
matter which variant is chosen, we have all entry points to call into.
The same entry point in different variants may have different target
requirements though.

The input to the linking stage are a collection of executables, each may
have multiple variants, but only ever provide one entry point. Together
with the above restriction, we can link two executables if and only if
their variants have the exact same set of target requirements. Under
such circumstances, we can make sure for a particular target requirement
(loaded as one variant during runtime), we can provide all entry points.