Standardizes CMake setup of C directory trees behind a macro. (#16011)
This was done 3-5 different ways in tree and even more out of tree. With
the new install/exports stuff being very sensitive to making sure that
everything gets exported properly, it becomes more imperative that this
always works more by default.

I needed to fix this for an out of tree project which needed to adapt to
export sets, so I just decided to do it nicely for everyone.

Apparently having these implicit deps as `iree_cc_library`'ies was
really fouling up compiler warning flag generation for unknown reasons.
We've definitely been running with degraded warnings. I fixed the things
that this uncovered.

I'm going to go ahead and land this because I have a broken downstream
and the implications of having these warning flags wrong is frightening.
Then I'll revert locally and try to root cause what mystical thing is
actually going on. Definitely don't want to accidentally get back into
this state.