Remove SYSTEM scope from transitive includes. (#16018)
Warnings should be fixed or handled as needed, not just by disabling
them for arbitrary portions of the codebase.

This was originally added to match Bazel, but Bazel builds are
undesirably more permissive because of it. This makes the CMake build
more correct.

We add an explicit SYSTEM_INCLUDES which preserves the old behavior and
use it for Tracy+related headers, which include things that are not
warning clean. Everything else is fine (after some of the newly
uncovered warnings were fixed).

Fixes #16017.