[cuda] Fix deadlock when advancing deferred queue in driver thread (#15673)
When GPU signals the CUevent-implemented semaphore to a new value, we
need to perform a `cuLaunchHostFunc()` to invoke the CPU to signal to
that value. At the same time, we can try to advance the deferred queue
to release more workload to the GPU. Advancing the deferred queue cannot
happen as part of the `cuLaunchHostFunc()` call, given that is in a
driver thread. Per the documentation:

  The host function must not make any CUDA API calls. Attempting
  to use a CUDA API may result in CUDA_ERROR_NOT_PERMITTED, but
  this is not required. The host function must not perform any
  synchronization that may depend on outstanding CUDA work not
  mandated to run earlier. Host functions without a mandated order
  (such as in independent streams) execute in undefined order and
  may be serialized.

We were actually doing that and seeing deadlock. So this commit changes
the implemention to create a new thread dedicated to monitoring and
releasing workload to the GPU, to avoid calling CUDA APIs as part of the
`cuLaunchHostFunc()` call.

Progress towards https://github.com/openxla/iree/issues/13245