e2e microkernel pipeline + argmax ukernel on ROCM backend. (#15943)
This commit presents is an end-to-end ukernel pipeline on ROCm with argmax as the first ukernel. It is
end-to-end in a because we automatically generate and link the bitcode, which means end users do not need to manually link bitcodes or hand-modify the kernel to have ukernel.generics.

This work is based on Ukernel lowerings on CPU written by Benoit Jacob
as well as a ukernel on CUDA sample by Mahesh Ravishankar.